Added video capture functionality for linux through VideoInputDevice and improved upon its performance.
Experimented with H.263+ encoding bit rate. The bit rate setting in the encoder context does not seem to be a literal bits/second setting. Rather, I've found the formula to be something like (although not perfect) bit rate = (bps/8) / 30 (for 30 frames per second). It appears that the bit rate setting is expecting bytes per frame, but I can't be sure. I need to look at the H.263 source code and see how it is being used.